pyproject.toml
[project]
name = "aws_chat"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Sehyeon", email = "raecrowned@gmail.com"},
    {name = "Kyuseok", email = "kyuseok000@gmail.com"}
]
dependencies = [
    "kafka-python>=2.0.2",
]
requires-python = "==3.8.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
distribution = false
